[ACTIVITI-3516] move conf package to right position
