+++ activiti-api-impl/activiti-api-process-runtime-impl/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-    org.conf.activiti.runtime.api.ConnectorsAutoConfiguration,\
-  org.conf.activiti.runtime.api.ProcessRuntimeAutoConfiguration,\
-  org.activiti.core.common.spring.connector.autoconfigure.ConnectorAutoConfiguration
+    org.activiti.runtime.api.conf.ConnectorsAutoConfiguration,\
+    org.activiti.runtime.api.conf.ProcessRuntimeAutoConfiguration,\
+    org.activiti.core.common.spring.connector.autoconfigure.ConnectorAutoConfiguration
